+++ build_runner_core/lib/src/generate/options.dart
@@ -134,6 +134,7 @@ class BuildFilter {
 /// Manages setting up consistent defaults for all options and build modes.
 class BuildOptions {
   final bool deleteFilesByDefault;
+  final bool delayWrites;
   final bool enableLowResourcesMode;
   final StreamSubscription logListener;
 
@@ -154,6 +155,7 @@ class BuildOptions {
   BuildOptions._({
     required this.debounceDelay,
     required this.deleteFilesByDefault,
+    required this.delayWrites,
     required this.enableLowResourcesMode,
     required this.logListener,
     required this.packageGraph,
@@ -172,6 +174,7 @@ class BuildOptions {
     LogSubscription logSubscription, {
     Duration debounceDelay = const Duration(milliseconds: 250),
     bool deleteFilesByDefault = false,
+    bool delayWrites = false,
     bool enableLowResourcesMode = false,
     required PackageGraph packageGraph,
     Map<String, BuildConfig> overrideBuildConfig = const {},
@@ -213,6 +216,7 @@ feature, you may need to run `dart run build_runner clean` and then rebuild.
     return BuildOptions._(
       debounceDelay: debounceDelay,
       deleteFilesByDefault: deleteFilesByDefault,
+      delayWrites: delayWrites,
       enableLowResourcesMode: enableLowResourcesMode,
       logListener: logSubscription.logListener,
       packageGraph: packageGraph,